Added available set names to Set node "set" plug popup menu.
Moved the functionality from SetFilterUI, generalised it to work with SceneProcessors as well as SceneFilters, and added a metadata item to trigger its operation. This will let us add the same menu to other nodes in the future (DeleteSets for instance).
